:sparkles: allowed file types and pairs

+++ backend/app/api/routes/files.py
@@ -5,16 +5,22 @@ from typing import Any
 
 from fastapi import APIRouter, Depends, HTTPException, Query, UploadFile, status
 from fastapi.responses import FileResponse
-from sqlalchemy import desc
+from sqlalchemy import and_, desc
 from sqlmodel import func, select
 
 from app.api.deps import CurrentUser, FileDep, SessionDep, get_current_user
 from app.core.config import settings
-from app.core.file_types import FileTypeMetadata, file_types
+from app.core.file_types import FileTypeEnum, FileTypeMetadata, file_types
 from app.core.security import create_access_token
 from app.crud import get_file_stats
 from app.crud import save_file as save_file_to_filesystem
-from app.models import File, FilePublic, FilesPublic, FilesStatistics, Message
+from app.models import (
+    File,
+    FilePublic,
+    FilesPublic,
+    FilesStatistics,
+    Message,
+)
 
 router = APIRouter()
 
@@ -24,13 +30,24 @@ def read_files(
     current_user: CurrentUser,
     skip: int = 0,
     limit: int = 100,
-    order_by: str = Query("-created_at", regex=r"^-?[a-zA-Z_]+$")
+    order_by: str = Query("-created_at", regex=r"^-?[a-zA-Z_]+$"),
+    types: list[FileTypeEnum] = Query(None),
 ) -> Any:
     """
     Retrieve saved files.
     """
+    if not types:
+        types = []
+    print(f"Types: {types}")
+    # start with all the top level files
+    base_where = (File.owner_id == current_user.id) & (File.saved) & (File.parent_id.is_(None))
+
+    # filter by file types if requested
+    if types:
+        base_where = and_(base_where, File.file_type.in_(t.value for t in types))
+
     # Counting for pagination
-    count_query = select(func.count()).select_from(File).where((File.owner_id == current_user.id) & (File.saved))
+    count_query = select(func.count()).select_from(File).where(base_where)
     count = session.exec(count_query).one()
 
     # Parse the order_by string to determine the column and direction
@@ -45,7 +62,7 @@ def read_files(
         raise HTTPException(status_code=400, detail=f"Invalid column name: {column_name}")
 
     # Build the query based on user role
-    query_base = select(File).where((File.owner_id == current_user.id) & (File.saved))
+    query_base = select(File).where(base_where)
 
     # Apply ordering, pagination and execute
     files_query = query_base.order_by(order_expression).offset(skip).limit(limit)
@@ -99,6 +116,46 @@ def upload_file(
 
     return file_metadata
 
+
+@router.post("/pairs", response_model=FilePublic)
+def create_pair(
+    session: SessionDep, current_user: CurrentUser, name: str, forward: uuid.UUID, reverse: uuid.UUID
+) -> Any:
+    """
+    Create a group of paired-end reads.
+    """
+    # Check if all files belong to the user
+    forward_file = session.get(File, forward)
+    reverse_file = session.get(File, reverse)
+    if not forward_file or not reverse_file:
+        raise HTTPException(status_code=404, detail="File not found")
+    if forward_file.owner_id != current_user.id or reverse_file.owner_id != current_user.id:
+        raise HTTPException(status_code=400, detail="Not enough permissions")
+
+    sum_size = forward_file.size + reverse_file.size
+    children = [forward_file, reverse_file]
+    # Create a pair group
+    pair_metadata = File(
+        name=name,
+        owner_id=current_user.id,
+        file_type=FileTypeEnum.PAIR.value,
+        children=children,
+        size=sum_size,
+        saved=True,
+    )
+    session.add(pair_metadata)
+    session.commit()
+    session.refresh(pair_metadata)
+    # set the created_at time for the children to the same as the parent
+    forward_file.created_at = pair_metadata.created_at
+    # add so reverse file is after forward file in the list R1 then R2
+    reverse_file.created_at = pair_metadata.created_at + timedelta(seconds=1)
+    session.add(forward_file)
+    session.add(reverse_file)
+    session.commit()
+    print(f"Pair created: {pair_metadata}")
+    return pair_metadata
+
 @router.get("/{id}", response_model=FilePublic)
 def read_file(session: SessionDep, current_user: CurrentUser, id: uuid.UUID) -> Any:
     """
@@ -140,12 +197,16 @@ def delete_file(session: SessionDep, current_user: CurrentUser, id: uuid.UUID) -
     if file_metadata.owner_id != current_user.id:
         raise HTTPException(status_code=400, detail="Not enough permissions")
     try:
-        # remove file
-        file_path = Path(file_metadata.location)
-        if file_path.exists():
-            file_path.unlink()
-        session.delete(file_metadata)
-        session.commit()
+        files_to_delete = [file_metadata]
+        if file_metadata.children:
+            files_to_delete.extend(file_metadata.children)
+        for file_to_delete in files_to_delete:
+            if file_to_delete.location:
+                file_path = Path(file_to_delete.location)
+                if file_path.exists():
+                    file_path.unlink()
+            session.delete(file_to_delete)
+            session.commit()
     except Exception:
         raise HTTPException(status_code=500, detail="Failed to delete file")
     return Message(message="File deleted successfully")
@@ -159,9 +220,10 @@ def delete_files(session: SessionDep, current_user: CurrentUser) -> Any:
     try:
         for file_metadata in files:
             # remove file
-            file_path = Path(file_metadata.location)
-            if file_path.exists():
-                file_path.unlink()
+            if file_metadata.location:
+                file_path = Path(file_metadata.location)
+                if file_path.exists():
+                    file_path.unlink()
             session.delete(file_metadata)
         session.commit()
     except Exception:
@@ -178,6 +240,11 @@ def download_file(session: SessionDep, current_user: CurrentUser, id: uuid.UUID)
         raise HTTPException(status_code=404, detail="File not found")
     if file_metadata.owner_id != current_user.id:
         raise HTTPException(status_code=400, detail="Not enough permissions")
+    if file_metadata.file_type == FileTypeEnum.PAIR.value:
+        # TODO: Download zipped files?
+        raise HTTPException(status_code=400, detail="Paired-end reads cannot be downloaded directly. Please download the files separately.")
+    if not file_metadata.location:
+        raise HTTPException(status_code=404, detail="File not found")
     return FileResponse(file_metadata.location, filename=file_metadata.name)
 
 @router.get("/{id}/token", response_model=str)
@@ -200,6 +267,8 @@ def download_file_with_token(file_metadata: FileDep) -> Any:
     """
     Download file by token.
     """
+    if not file_metadata.location:
+        raise HTTPException(status_code=404, detail="File not found")
     file_path = Path(file_metadata.location)
     if not file_path.exists():
         raise HTTPException(status_code=404, detail="File not found")
